pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supwngo"
version = "1.0.0"
description = "SupwnGo - Automated binary exploitation framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "SupwnGo Contributors", email = "supwngo@example.com"}
]
keywords = [
    "pwn",
    "exploit",
    "binary",
    "security",
    "ctf",
    "rop",
    "shellcode",
    "fuzzing",
    "vulnerability",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Software Development :: Disassemblers",
]
dependencies = [
    "pwntools>=4.11.0",
    "angr>=9.2.0",
    "angrop>=9.2.0",
    "capstone>=5.0.0",
    "keystone-engine>=0.9.2",
    "ropper>=1.13.0",
    "pyelftools>=0.29",
    "click>=8.1.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.0",
    "requests>=2.28.0",
    "networkx>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=1.0.0",
]

[project.scripts]
supwngo = "supwngo.cli:main"

[project.urls]
Homepage = "https://github.com/jeremylaratro/supwngo"
Documentation = "https://github.com/jeremylaratro/supwngo#readme"
Repository = "https://github.com/jeremylaratro/supwngo"
Issues = "https://github.com/jeremylaratro/supwngo/issues"

[tool.setuptools.packages.find]
exclude = ["tests*", "build*"]

[tool.setuptools.package-data]
supwngo = ["payloads/templates/*.py", "data/*.json"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["supwngo"]
branch = true
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
